Rename sheet_activity to activities
