+++ package/lib/smooth_sheets.dart
@@ -10,7 +10,7 @@ export 'src/foundation/framework.dart';
 export 'src/foundation/keyboard_dismissible.dart';
 export 'src/foundation/notifications.dart';
 export 'src/foundation/physics.dart';
-export 'src/foundation/sheet_activity.dart';
+export 'src/foundation/activities.dart';
 export 'src/foundation/sheet_content_scaffold.dart';
 export 'src/foundation/sheet_controller.dart' hide SheetControllerScope;
 export 'src/foundation/sheet_extent.dart';
