+++ package/lib/src/foundation/sheet_extent.dart
@@ -5,7 +5,7 @@ import '../internal/double_utils.dart';
 import '../scrollable/scrollable_sheet.dart';
 import '../scrollable/scrollable_sheet_extent.dart';
 import 'physics.dart';
-import 'sheet_activity.dart';
+import 'activities.dart';
 import 'sheet_controller.dart';
 import 'sheet_status.dart';
