+++ package/lib/src/navigation/navigation_route.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import '../foundation/framework.dart';
-import '../foundation/sheet_activity.dart';
+import '../foundation/activities.dart';
 import '../foundation/sheet_extent.dart';
 import '../foundation/sheet_status.dart';
 import 'navigation_sheet.dart';
