+++ package/lib/src/navigation/navigation_sheet.dart
@@ -6,7 +6,7 @@ import 'package:flutter/widgets.dart';
 import '../foundation/framework.dart';
 import '../foundation/keyboard_dismissible.dart';
 import '../foundation/physics.dart';
-import '../foundation/sheet_activity.dart';
+import '../foundation/activities.dart';
 import '../foundation/sheet_controller.dart';
 import '../foundation/sheet_extent.dart';
 import '../foundation/sheet_status.dart';
